Changed the setIdleTimeout method to take an int instead of a long. The socket interface takes the so_timeout as an int and also the AMQP heartbeat interval is taken as an int. This change will help simplify the code that handles heartbeats.
git-svn-id: https://svn.apache.org/repos/asf/qpid/trunk/qpid@901000 13f79535-47bb-0310-9956-ffa450edef68

+++ java/client/src/main/java/org/apache/qpid/client/AMQConnectionDelegate_0_10.java
@@ -295,7 +295,7 @@ public class AMQConnectionDelegate_0_10 implements AMQConnectionDelegate, Connec
 
     public void setIdleTimeout(long l)
     {
-        _qpidConnection.setIdleTimeout(l);
+        _qpidConnection.setIdleTimeout((int)l);
     }
 
     public int getMaxChannelID()
